Port pytools-based analysis to HL7800
The pytools-based elfsize.py analysis use on ThreadX products uses
debug info to trace symbols back to their source, so can provide a
more accurate view of memory usage than the previous GNU Toolchain
based solution.  So Port this solution to to HL78 MAP, and use it
as the default size analysis.  Keep the old analysis engine as
alt1250-map-gnu for comparison purposes.

Resolves: LE-15531
Change-Id: If4998ec4f32d6b916e5f53f72f7d562268690527